[stabilizer][zmpcc] Remove unused ZMPCC
- As suggested by @caron, ZMPCC is mostly unused:

"I suggest we remove this one as we now set the CoM admittance to zero most of the time.
Historically, it has helped us (1) improve sagittal sway on HRP-4 while stair climbing (September 2018) and (2) improve lateral stepping with 0.7/0.1 phase durations. In my view, it only helps when the ZMP saturates the support area, which hopefully does not happen with a properly tuned walking pattern 😉
The drawback of ZMPCC is that it may pick up vibrations, which is why we chose to disable it in Saint-Nazaire (February 2019). Actually, we still had vibrations afterwards 😅 and didn't validate that they were due to ZMPCC (hence the chest_massage plans and other fond memories). But the main thing is that the robot worked essentially the same afterwards. I just did some simulations on HRP-2Kai and confirmed that we only see this drawback in nominal walking.
So, in the end I would say: let's make the LIPMStabilizerTask as simple as possible and remove ZMPCC."