[mc_rbdyn] Introduce FreeFrame
FreeFrame is not attached to a robot, for now it only defines two
signals: position and velocity whose values can be directly set by the
user.

A FreeFrame can also be defined relative to another FreeFrame and this
relative transformation can be manipulated by the user as well. In such
cases, the absolute position/velocity of the frame cannot be set
anymore.

Furthermore:
- Frame derives from FreeFrame
- Frame therefore correctly honor the hierarchy (before all frames were
directly attached to their parent body)
- Frame transformation from the parent frame to the frame can be redefined

ToDo:
- Add acceleration signal
- Introduce all signals from Frame into FreeFrame (disabled in FreeFrame but enabled in Frame)
- Allow the use of FreeFrame in some mc_tvm functions (e.g. to provide objectives)